feat: Move pr-ci scripts to be executable with uv
The pr-ci scripts were refactored to be directly executable with `uv
run`. This change removes the need for a separate `pyproject.toml` and
`uv.lock` file, simplifying dependency management and allowing for
easier execution of the scripts. The internal imports were updated to
reflect the new file structure.
