+++ .tekton/issue-create.yaml
@@ -81,7 +81,7 @@ spec:
                 - name: GEMINI_MODEL
                   value: "$(params.gemini_model)"
               script: |
-                uv run --directory ./hack/pr-ci pr-ci issue-create
+                uv run --directory ./hack/pr-ci main.py issue-create
         params:
           - name: repo_url
             value: "$(params.repo_url)"
