+++ .tekton/jira.yaml
@@ -110,7 +110,7 @@ spec:
                 - name: JIRA_ISSUETYPE
                   value: "$(params.jira_issuetype)"
               script: |
-                uv run --directory ./hack/pr-ci pr-ci jira-create
+                uv run --directory ./hack/pr-ci main.py jira-create
         params:
           - name: repo_url
             value: "$(params.repo_url)"
