+++ .tekton/pr-ci.yaml
@@ -87,7 +87,7 @@ spec:
                 - name: PR_NUMBER
                   value: "$(params.pull_request_number)"
               script: |
-                uv run --directory ./hack/pr-ci pr-ci lint
+                uv run --directory ./hack/pr-ci main.py lint
             - name: analyze-and-label
               image: ghcr.io/astral-sh/uv:debian
               workingDir: $(workspaces.source.path)
@@ -117,7 +117,7 @@ spec:
                 - name: MAX_LABELS
                   value: "$(params.max_labels)"
               script: |
-                uv run --directory ./hack/pr-ci pr-ci update
+                uv run --directory ./hack/pr-ci main.py update
         params:
           - name: repo_url
             value: "$(params.repo_url)"
