--- hack/pr-ci/pyproject.toml
@@ -1,21 +1,0 @@
-[project]
-name = "pr-ci"
-version = "0.1.0"
-requires-python = ">=3.11"
-dependencies = [
-  "requests",
-  "google-generativeai",
-]
-
-[project.scripts]
-pr-ci = "pr_ci.main:main"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/pr_ci"]
-
-[tool.uv]
-package = true
